+++ packages/language/src/snakeskin-module.ts
@@ -5,6 +5,9 @@ import { SnakeskinValidator, registerValidationChecks } from './snakeskin-valida
 import { SnakeskinTokenBuilder, SnakeskinLexer } from './parser/snakeskin-lexer.js';
 import { SemanticTokenProvider } from './semantic-tokens.js';
 import { HoverProvider } from './hover-provider.js';
+import { TypeScriptServices } from './typescript-service.js';
+import { SnakeskinDefinitionProvider } from './definition-provider.js';
+import { SnakeskinWorkspaceManager } from './workspace-manager.js';
 
 /**
  * Declaration of custom services - add your own service classes here.
@@ -12,14 +15,21 @@ import { HoverProvider } from './hover-provider.js';
 export type SnakeskinAddedServices = {
     validation: {
         SnakeskinValidator: SnakeskinValidator
-    }
+    },
+    TypeScript: TypeScriptServices,
 }
+
+export type SnakeskinSharedServices = {
+    workspace: {
+        WorkspaceManager: SnakeskinWorkspaceManager;
+    };
+};
 
 /**
  * Union of Langium default services and your custom services - use this as constructor parameter
  * of custom service classes.
  */
-export type SnakeskinServices = LangiumServices & SnakeskinAddedServices
+export type SnakeskinServices = LangiumServices & SnakeskinAddedServices & { shared: SnakeskinSharedServices };
 
 /**
  * Dependency injection module that overrides Langium default services and contributes the
@@ -32,11 +42,19 @@ export const SnakeskinModule: Module<SnakeskinServices, PartialLangiumServices &
         Lexer: (services) => new SnakeskinLexer(services),
     },
     validation: {
-        SnakeskinValidator: () => new SnakeskinValidator(),
+        SnakeskinValidator: (services) => new SnakeskinValidator(services),
     },
     lsp: {
         SemanticTokenProvider: (services) => new SemanticTokenProvider(services),
         HoverProvider: (services) => new HoverProvider(services),
+        DefinitionProvider: (services) => new SnakeskinDefinitionProvider(services),
+    },
+    TypeScript: (services) => new TypeScriptServices(services),
+};
+
+export const SnakeskinSharedModule: Module<LangiumSharedServices, SnakeskinSharedServices> = {
+    workspace: {
+        WorkspaceManager: (services) => new SnakeskinWorkspaceManager(services),
     },
 };
 
@@ -61,7 +79,8 @@ export function createSnakeskinServices(context: DefaultSharedModuleContext): {
 } {
     const shared = inject(
         createDefaultSharedModule(context),
-        SnakeskinGeneratedSharedModule
+        SnakeskinGeneratedSharedModule,
+        SnakeskinSharedModule,
     );
     const Snakeskin = inject(
         createDefaultModule({ shared }),