+++ packages/extension/CHANGELOG.md
@@ -10,11 +10,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - CLI subcommand for running validation checks.
+- Go to definition support for `include` directives.
+- Validation that `include` directive paths exist.
 
 ### Fixed
 
 - Some diagnostic messages related to parsing errors around indentation were not being shown.
-- Parser now allows commas to appear at the beginning of plain text
+- Parser now allows commas to appear at the beginning of plain text.
 
 ## [0.0.4] - 2024-08-15